Update SPI driver to reflect register changes
CS lines are now controlled by a register in the SPI
controller itself.
FIFO depths are now available in the 'info' register, and
the FIFOs are reduced to 16 entries apiece.

Ensure that zero-byte transfer requests do not result in a
deadlock between software and hardware; the controller will
attempt to start but cannot perform a zero-byte transfer.

Zero-byte blocking_write/read may be used instead to
synchronise with the controller core.