Add SPI internal loopback and software reset
Internal loopback within the SPI controller is trivial and very
useful for testing; simultaneous transmit and receive shall result
in all transmitted bytes being replicated exactly in the receive
FIFO; we can test permutations of polarity and phase settings.

Software reset is required to recover in the event of an error or
indeed just a re-download of the software without an intervening
IP block reset presently. Clear the TX FIFO, reset the controller
core and then clear the RX FIFO.