Refactor remainder of USB Device driver
This commit performs the rest of the refactoring for the USB device
driver. This primarily involves renaming variables/attributes/methods
etc. to remove abbreviations, or extracting guard clauses to reduce the
amount of nesting within functions. `clang-format` and `clang-tidy` are
also used at this stage to enforce the CHERIoT RTOS code standards. A
bit more documentation is also added.

Apart from the names of functions/attributes, the only other public
change to the interface is to the `packet_receive` function, which has
had its first two parameters (buffer ID and endpoint ID) swapped around,
so that its interface is consistent with the interface of the
`packet_send` function.