Update i2c blocking write to block, and check for NACKs
Blocking write currently blocks only when the FIFO is full. This makes
it block till the FIFO is actually emptied, and checks in the write
whether the controller FSM halted (e.g. due to a NACK), and clears it if
so. This allows the I2C `blocking_write` function to be reliably used
and stops it hanging upon a write failure when e.g. the device doesn't
exist. Without this, the write will appear to be successful but then a
`blocking_read` will just halt indefinitely as the format FIFO will
never be empty, which is not ideal. Now you know when calling
`blocking_write` whether it failed or not, and can accordingly choose
whether to call `blocking_read`.

Shouldn't impact performance much as FIFOs are quite small anyhow.