Update Sonata PWM driver to fix issues and match latest HW
This commit updates Sonata's PWM driver in a couple of ways. Firstly, it
makes sure that the number of PWM instances is correct and matches the
number of PWM outputs that are actually available in the harware.

Second, it separates out the number of instances, to provide separate
driver interfaces for the general (muxable) PWM and the LCD's PWM used
for its backlight.

It also makes a small fix, moving the debug assert to a regular check
(and having the `output_set` function return whether it was successful
or not), to ensure that the instance check is applied even if debug
logging is disabled. It removes the check that the duty cycle is less
than the period, because on Sonata's RTL, having duty cycle > period is
the only way to get a 100% high output over PWM, which is useful.