Examples now have model output to check against
This is useful in CI to see if examples are running correctly. The
output is added to the scripts directory so that it is not confusing to
users that are not necessarily familiar with CI. The names of the
examples do not include the number because at the moment I've not found
a clean way to get the number in xmake because the firmware image has
the number removed. Also, output may vary per board so the model output
is first divided up per board.